Move contexts to own file not included by worker
