+++ src/components/ViewerPanel.tsx
@@ -1,7 +1,7 @@
 // Portions of this file are Copyright 2021 Google LLC, and licensed under GPL2+. See COPYING.
 
 import React, { CSSProperties, forwardRef, useContext, useEffect, useRef, useState } from 'react';
-import { ModelContext } from '../state/model';
+import { ModelContext } from './contexts';
 import { StlViewer} from "react-stl-viewer";
 
 export default function ViewerPanel({className, style}: {className?: string, style?: CSSProperties}) {
