+++ src/components/contexts.ts
@@ -1,0 +1,7 @@
+import React from "react";
+import { Model } from "../state/model";
+
+export const FSContext = React.createContext<FS | undefined>(undefined);
+
+export const ModelContext = React.createContext<Model | null>(null);
+
