+++ src/state/model.ts
@@ -1,6 +1,5 @@
 // Portions of this file are Copyright 2021 Google LLC, and licensed under GPL2+. See COPYING.
 
-import React from "react"
 import { checkSyntax, render, RenderArgs, RenderOutput } from "../runner/actions";
 import { MultiLayoutComponentId, SingleLayoutComponentId, State } from "./app-state";
 import { bubbleUpDeepMutations } from "./deep-mutate";
@@ -163,6 +162,3 @@ export class Model {
     }})
   }
 }
-
-export const ModelContext = React.createContext<Model | null>(null);
-
